EntityId parser, also move entities parser to its own file
